refactor: card css 파일 분리 및 text-ellipsis 및 글자 크기 수정
